switch threshold to use uint32

+++ internal/graph/dispatch_throttling_check_resolver.go
@@ -104,9 +104,9 @@ func (r *DispatchThrottlingCheckResolver) ResolveCheck(ctx context.Context,
 
 	thresholdInContext := ctx.Value(telemetry.DispatchThrottlingThreshold)
 	if thresholdInContext != nil {
-		thresholdInInt, ok := thresholdInContext.(int)
+		thresholdInInt, ok := thresholdInContext.(uint32)
 		if ok {
-			threshold = min(uint32(thresholdInInt), r.config.MaxThreshold)
+			threshold = min(thresholdInInt, r.config.MaxThreshold)
 		}
 	}
